+++ packages/lib/dawproject/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.76](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-dawproject@0.0.75...@opendaw/lib-dawproject@0.0.76) (2026-08-17)
+
+**Note:** Version bump only for package @opendaw/lib-dawproject
+
 ## [0.0.75](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-dawproject@0.0.74...@opendaw/lib-dawproject@0.0.75) (2026-08-11)
 
 **Note:** Version bump only for package @opendaw/lib-dawproject