+++ packages/lib/fusion/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.100](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-fusion@0.0.99...@opendaw/lib-fusion@0.0.100) (2026-08-17)
+
+**Note:** Version bump only for package @opendaw/lib-fusion
+
 ## [0.0.99](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-fusion@0.0.98...@opendaw/lib-fusion@0.0.99) (2026-07-16)
 
 **Note:** Version bump only for package @opendaw/lib-fusion