+++ packages/studio/adapters/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-adapters@0.2.1...@opendaw/studio-adapters@0.3.0) (2026-08-17)
+
+### Features
+
+- **icons:** give the Vaporisateur its own icon ([1dbad8e](https://github.com/andremichelle/openDAW/commit/1dbad8e6aa2c8835877300a39fc45af2ce272ffb))
+
 ## [0.2.1](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-adapters@0.1.6...@opendaw/studio-adapters@0.2.1) (2026-08-11)
 
 **Note:** Version bump only for package @opendaw/studio-adapters