+++ packages/studio/core-wasm/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.13](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core-wasm@0.0.12...@opendaw/studio-core-wasm@0.0.13) (2026-08-17)
+
+**Note:** Version bump only for package @opendaw/studio-core-wasm
+
 ## [0.0.12](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core-wasm@0.0.10...@opendaw/studio-core-wasm@0.0.12) (2026-08-11)
 
 **Note:** Version bump only for package @opendaw/studio-core-wasm